Added files used for project
This is my first java commit, so I'm not entirely sure everything that I'm supposed to upload. Also, since this was an education project, rather than a personal one, the folders are all a bit wonky in the way that a lot of them contain things such as test cases.